Implement extensible pagination
Create Pagination.tsx, which should be able to be slotted into any page that has a lengthy list of child elements. This commit enables pagination for the EventsPage.tsx, but this should prove suitable for other pages as well. Of note: this method of pagination does not reduce the initial load time, just changes the way elements are displayed, potentially making them easier to comb through.