Improve caption for Gamma 3/4
Gammas 3 and 4 lasted for two weeks. For all other seasons, the end date is
calculated by adding 6 days to whatever the first day was. Since this doesn't
work for the given seasons, and since FeedSeasonList says that gamma9 had _two_
seasons in it (and doesn't mention gamma10 at all), there is now a helper
function which hardcodes in these two edgecases.

Hopefully TGB will update FeedSeasonList when they bring the feed back in
the next era, and that will actually contain the correct values for seasons.