Extend shorthand for single season SCs
Create a helper method to format sim/season and sim/season/day for use
wherever that data is displayed.

Unfortunately only returns plaintext, so not a fix for the style mismatch
on the event page -- all of "Gamma X" is bolded, which does not match the
case for sim=THISIDISSTATICYO where only the season number has emphasis.