Add support for new games
SUPER BROKEN

The way that I'm combining state in this is very much incorrect. It's duplicating
entries/steps, and I'm not sure why.

Additionally, the model structure may not be right -- the baserunners isn't ever
getting written to, and the balls jump around all over the place.

AND, the innings seem to switch at random points? Not sure what that's about.

I am confident on fixing the timestamp (or at least, some of the timestamps).
Currently, we're only passing around the internal blaseball data, but we need to
instead pass around the "ChronEntityVersionExperimental" into most places,
which will let us use the timestamp in there.

Also some of the displayText values are empty?