Add brand new hack fix
If the game is not being filtered, the firstUpdate in a group will not have a
pitcher (presumably because they haven't stepped up to the plate) (this
is fine if you're only showing important games because important events
can't happen _until_ the pitcher is in position, I guess).

This patch hacks things to use the second event in an inning. This is
almost certainly going to be a falsehood at some point, but "at some
point" is not now.

(You can't use the second event when we're only showing important
events, because there might only be one important event in an inning.
Good job lofty for testing before deploying).